Clean up release_commands.py and make output pretty
Instead of two separate release quad for loops for apps and samples, the
release process is now split into several stages:
- Figure out what "targets" do we need to build by parsing east.yml and
  change them into jobs.
- Run these jobs one after another
- Create zip packages afterwards.

The stdout is dramatically changed, by default we do not show the
west/Cmake output but hide it behind progress bars.
If something goes wrong we abort and show the build output to the user.
User can turn on the output with --verbose flag.