Simplify east build, making it a almost true wrapper command
east build does not anymore exposes specific west options, it just adds
--build-type functionality over it.

As the Click command parser (actually a copy of optparse) silently drops
double-dash "--" the task of determining if a positional argument
(source_dir) was passed is much harder.

This was done by making a copy of raw args before they were parsed and
working with them, instead with args received by the build command.

create_build_command function still needs to be clean up but this will
need to wait for now.