Correctly handle ctrl+c when running commands
It was noted that some programs do not receive the SIGINT when ctrl+c is pressed,
specifically, codechecker would keep running until it does its things.
To fix this case, the SIGINT is explicitly passes to the created
subprocess.

This commit also fixes a long standing issue with running gdb commands
(via east debug). SIGINT can be ignored by setting a specific flag, in
that case gdb correctly handles the SIGINT by itself.