Break east update api, remove east sys-setup and add east install
This is a breaking change.

east update command does not update toolchain anymore, it is just a
wrapper for the west update command.

east sys-setup command was completely removed and replaced by east
install command, which also contains logic to install toolchains.

Other than that most of the changes in this commit are either:
* moving code around
* renaming things to better express their new purpose (package -> tool)
* correcting messages