Trully trully fix the single app in east.yml in multi app repo
I though that I fixed this in 1054605. And then I though that I fixed
the issue in 7fad75e. But I didn't.

The problem was present in the repos where only one app (for example
app_one) was specified in the east.yml, but the actual "app" folder
would contain two apps (for example app_one and app_two).

When running east release, app_one would not get correct src_dir for the
build command, src_dir would be just "app", instead of "app/app_one".

As it can be seen in the changes made in the release_commands.py file
some heuristic check involving CMakeLists.txt had to be made to truly
determine if this is a single app repo or not.

This issue was ridiculous. The problem space that east.yml needs to solve
is too big.

Related: #85 (first task)