Add back cppcheck to the codechecker install list
Although cppcheck is not used, the codechecker --help command (and a lot
of other commands) fails a version check if cppcheck binary is not
present.

I am currently assuming that this is due to the contents of
codechecker/build/CodeChecker/config/package_layout.json that I modify
to tell codechecker where are analyzer binaries. Maybe a better fix
would be to not write anything for the cppcheck, however I want to leave
this inside until some point in the future where I have time to properly
configure cppcheck for our use.