Fix the issue with -fno-printf-return-value
east codechecker check command would fail due to this flag, as the flags
comes from the GCC, but clang doesn't know about it.

There is no way to ignore this on the clang level, but for some reason
clangd knows how to ignore this. So the only solution is to remove this
from the compile_commands.json

There are many issues on the Zephyr about this (as Picolib introducted
this), but there is no general solution for this yet.

Related:
https://github.com/zephyrproject-rtos/zephyr/issues/68831#issuecomment-1937007486