feat: migrate to using toolchain-manager with nrfutil
Instead of downloading the nrfutil-toolchain-manager.exe executable from
a Nordic's project on the GitHub, use nRF Util.

nRF Util supports the toolchain-manager command for quite a while now
(more than a year). Due to the few issues that came up recently (macOS
support and v2.7.0 NCS toolchain support) it was quite the time to
perform the migration.

IMPORTANT: this changes were for now tested only on the Windows.