feat: make east compatible with docker
When inside the docker environment East will not try to pass any
commands to the nRFUtil's toolchain-manager, but it will pass them
directly to the west.

The rationale for that is that the docker environment should provide
all commands that are needed for working on a Zephyr/NCS project and
east should just use them directly.