feat: add east util version command
This command generates VERSION files on the paths that are either given
in the east.yml file or on the command line.

The versioning information is parsed from the output of the git describe
command (with the help of the ParsedTag and ZephyrSemver classes),
however user can also provide that info on the command line.

Related: https://github.com/IRNAS/irnas-zephyr-template/issues/29#issuecomment-2307001432