feat: introduce EAST_NRFUTIL_CI_MODE, change conf of nrfutil
EAST_NRFUTIL_CI_MODE env var replaced the
EAST_DONT_USE_TOOLCHAIN_MANAGER. Latter had a bit confusing name and
buggy behaviour in some instances.

When EAST_NRFUTIL_CI_MODE is set, East will not download the nrfutil or
try to configure it. It is assumed that the environment provides
correctly confgured nrfutil on PATH.

Additionally, East will assume that it doesn't need to run shell commands
through the `nrfutil toolchain-manager launch` command to get access to
various toolchain commands (such as `python`, C compilers, etc.).
It is expected that all commands are available on the `PATH`.

For the normal non-ci mode the configuration of the nrfutil changed.
Whenever the east install nrfutil-toolchain-manager is called, the
nrfutil, device command and toolchain-manager command will be upgraded
to the pinned versions and toolchain install location will be
configured.

Previously the version updates were only happening if the nrfutil wasn't
installed, thus leaving the space for possible version mismatches.