fix: change deprecation warnings and pin dependencies
This is more of a bandaid solution to a serious problem that I don't yet
know how to properly resolve.

In the newer ncs-zephyr docker images where east is used we noticed that
east wouldn't run at all, due to the click package complaining that it
received unexpected `deprecation` parameter.

After some using the pip freeze to check the installed dependencies in
the docker image it was discovered that pip was installing an older
version of the click package (8.1.8), instead of the one that was
specified by the east (8.2.1). The older version didn't yet use
`deprecation` parameter (or it didn't use it in the same way, I didn't
dig too deep into this).

After learning about this, I wanted to know why that is the case, so,
why pip wasn't respecting the values in the scripts/requirements.txt.

It turns out that using dynamic requirements really doesn't mean what I
though it does, in the end only the names of the package get into the
required section of the east's wheel metadata, while the exact versions
are dropped.

Since the pip in the case of the docker didn't saw that, it just
installed whatever version of the rich that was specified by some other
python package in the docker environment context.

Now, correct was to resolve this would be to change how east-tool is
installed in the container, so instead of using pip, the pipx should be
used. But after several tries I gave up on this. Due to the weird
modification of the python environment the pipx doesn't really work as
it should.

So what I did? I removed the `deprecation` parameter, added plain
deprecation warnings and explicitly pinned the dependencies, making sure
that I pin an older version of the click package. I then tested this
kind of east install inside the docker to confirm that everything works.

This change won't impact the east users that use it outside the docker.
There the pipx is used and everything is ok. But inside the docker
container we are basically at the mercy with the requirements that
Nordic toolchain manager specifies... Not ideal.